+++ docs/changes.rst
@@ -21,6 +21,21 @@ To be released.
     :mod:`xml.etree.ElementTree` can recognize when finding children elements.
 
 
+Version 0.3.2
+-------------
+
+Released on November 5, 2014.
+
+- Fixed a bug that :class:`~libearth.subscribe.SubscriptionList`\ s having
+  :class:`~libearth.subscribe.Outline`\ s without their :attr:`created_at
+  <libearth.subscribe.Outline.created_at>` attribute failed to be merged
+  on Python 3.  [:issue:`65`]
+- Fixed a bug that a :class:`~libearth.schema.DocumentElement` in streamed
+  read mode is not properly marked as complete even when it's completed by
+  :func:`~libearth.schema.complete()` function in some cases.
+>>>>>>> 0.3.2
+
+
 Version 0.3.1
 -------------